# Refactor `scatter.cuh` and `gather.cuh`