--- a/cpp/src/partitioning/partitioning.cu
+++ b/cpp/src/partitioning/partitioning.cu
@@ -433,14 +433,22 @@ struct copy_block_partitions_dispatcher {
                                          stream);
 
     // Use gather instead for non-fixed width types
-    return type_dispatcher(input.type(),
-                           detail::column_gatherer{},
-                           input,
-                           gather_map.begin(),
-                           gather_map.end(),
-                           false,
-                           stream,
-                           mr);
+    //    return type_dispatcher(input.type(),
+    //                           detail::column_gatherer{},
+    //                           input,
+    //                           gather_map.begin(),
+    //                           gather_map.end(),
+    //                           false,
+    //                           stream,
+    //                           mr);
+    return std::move(gather(table_view{{input}},
+                            gather_map.begin(),
+                            gather_map.end(),
+                            out_of_bounds_policy::DONT_CHECK,
+                            stream,
+                            mr)
+                       ->release()
+                       .front());
   }
 };