… cpp/include/cudf/detail/scatter.cuh
… cpp/src/copying/scatter.cu
… cpp/src/partitioning/round_robin.cu
… cpp/src/quantiles/quantiles.cu
… cpp/src/reshape/tile.cu