This PR refactors `scatter.cuh` and `gather.cuh`, totally moving all the implementations of the templated functions `detail::scatter` and `detail::gather` into separates `*.cu` files with explicit instantiations.

Typically, this is not a trivial task since the APIs `scatter` and `gather` take templated iterator parameters (iterators for scatter/gather maps), and the types of those iterators can be very complex templated `thrust::transform_iterator`. In this PR, in order to explicitly instantiate the `scatter` and `gather` APIs, I have to limit the iterators types to be only pointers and `thrust::constant_iterator`. For `thrust::transform_iterator`, I create a temporary `device_uvector` and fill it with the transformed iterator values then pass its `begin()/end()` iterators to the `scatter` and `gather` APIs. Since most cudf APIs are passing scatter/gather maps from a `device_uvector` iterators, I don't think this would incur much overhead. The files affected by this are here: https://github.com/rapidsai/cudf/pull/8464/commits/37366cfcd2e3cac8b89ac751137d8898fd28758c.

After doing this (rough) refactoring, I gained about 5% speedup in compile time (on a 18 cores machine, 64 compile threads):
Before:
```
cudf: 6:48s
Others: 7:19s
All: 13:13s
```
After:
```
cudf: 6:31s
Others: 7:19s
All: 12:55s
```

It seems that 5% is not much worth do this, but if we merge this (after more cleanup) we will have a very nice scatter and gather APIs implementations: the header files `*.cuh` contain only declarations, and the implementations will be freed from including into many many others `*cu` file. Thus, I need more opinions from you guys about this before spending more time working on it.

-----------
### PS: This is not a complete work, just a rough draft and still needs much cleanup.