@jrhemstad 2021-06-09
I am not a fan of requiring materializing any iterator based gather or scatter map into a `device_uvector`.