@harrism 2021-06-09
I don't like the idea of forcing materialization. That kills the benefit of an iterator interface.

`cudf::reverse` is an example of a case where we don't need to materialize the gather map. https://github.com/rapidsai/cudf/pull/8410