@ttnghia 2021-06-09
> I don't like the idea of forcing materialization. That kills the benefit of an iterator interface.
> 
> `cudf::reverse` is an example of a case where we don't need to materialize the gather map. #8410

Yeap, this is a trade-off.

Based on the opinions, it seems that we should not continue doing this...